chair, table 컴포넌트분리
