+++ src/pages/LayoutSettingPage/components/GridBackground.tsx
@@ -9,6 +9,10 @@ import type {
 } from 'react';
 import type { Layout } from 'react-grid-layout';
 
+import {
+  ChairComponent,
+  TableComponent,
+} from 'pages/LayoutSettingPage/components/Chair';
 import { Popover } from 'pages/LayoutSettingPage/components/Popover';
 import { PopoverGpt } from 'pages/LayoutSettingPage/components/PopoverGpt';
 import {
@@ -71,11 +75,6 @@ export const GridBackground: React.FC<GridBackgroundProps> = ({
     saveLayoutChange(layout);
   };
 
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const handleTogglePopover = () => {
-    setIsPopoverOpen(!isPopoverOpen);
-  };
-
   return (
     <ShopGridBackground
       layout={myLayout}
@@ -108,27 +107,6 @@ export const GridBackground: React.FC<GridBackgroundProps> = ({
   );
 };
 
-interface ChairComponentProps extends ComponentPropsWithRef<'div'> {
-  isClickable: boolean;
-}
-// Chair Component
-export const ChairComponent = forwardRef(
-  ({ isClickable, ...rest }: ChairComponentProps, ref: Ref<HTMLDivElement>) => (
-    <ChairBorder {...rest} ref={ref}>
-      <Chair isClickable={isClickable} />
-    </ChairBorder>
-  ),
-);
-
-// Table Component
-export const TableComponent = forwardRef(
-  ({ isClickable, ...rest }: ChairComponentProps, ref: Ref<HTMLDivElement>) => (
-    <div>
-      <GridTable isClickable={isClickable} {...rest} ref={ref} />
-    </div>
-  ),
-);
-
 const ShopGridBackground = styled(GridLayout)<{
   width: number;
   $height?: number;
@@ -138,33 +116,4 @@ const ShopGridBackground = styled(GridLayout)<{
   }};
 
   height: ${({ $height }) => $height + 'px'};
-`;
-
-export const GridTable = styled.div<{ isClickable: boolean }>`
-  position: relative;
-  background-color: ${(props): string => props.theme.palette.grey[100]};
-
-  border-color: ${({ theme }) => theme.palette.black.main};
-  border-width: 1px;
-  border-style: solid;
-
-  cursor: ${({ isClickable }) => (isClickable ? 'pointer' : 'default')};
-`;
-
-// 의자 바깥에 투명한 테두리를 넣기 위함
-export const ChairBorder = styled.div`
-  ${flexSet()}
-`;
-
-// 검정 테두리를 준 의자 영역
-export const Chair = styled.div<{ isClickable: boolean }>`
-  background-color: ${(props): string => props.theme.palette.grey[100]};
-
-  width: ${CHAIR_SIZE_PX - CHAIR_BORDER_PX}px;
-  height: ${CHAIR_SIZE_PX - CHAIR_BORDER_PX}px;
-
-  border: ${CHAIR_BORDER_PX}px solid ${({ theme }) => theme.palette.black.main};
-  border-radius: 50%;
-
-  cursor: ${({ isClickable }) => (isClickable ? 'pointer' : 'default')};
 `;
